HW_models: NHW_NVMC: fix buffer read validation
The address validation in `nhw_nmvc_read_buffer` should only fail if
the address goes past the end of flash, not to the very end of flash.
For example, a 1 byte read at address `0x7FFFF` should be fine, but the
previous check would fail.

This affects trying to use the storage partition as a NVS backend.

Signed-off-by: Jordan Yates <jordan@embeint.com>