54L15: Change CLOKPOWER IRQ line (MDK 8.68, nrfx 3.9.0)
WARNING! This change breaks backwards compatibility with SW
which expected this line to be 270 for 54L15 devices,
and therefore with MDKs < 8.68 & nrfx < 3.9.0.
This change requires updating to the MDK >=8.68 and nrfx >=3.9.0
if building using the 54L15 models.
52 and 53 models are not affected.

Change the CLOCK_POWER interrupt line from 270 to 261.
This has changed in the latest MDK (8.68) and the new
value is used in the latest nrfx (3.9.0) drivers.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>