54 CCM: Fix UBSAN reported issue with zero size VLAs
With empty packets or with missconfigured joblists
three VLAs could have size 0.
A VLA with size 0 is in principle not C99 valid code and UBSAN complains
about it.
And for `a_copy[]` we unconditionally access byte 0, pressuming the joblist
is not missconfigured.
So let's allocate all of them with 1 extra byte which is a trivial stack
"waste" to avoid the issue.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>