GRTC: With a CC in the past evaluate immediately
Some SW relies on writing a past value to a CC, and having that triger
right away an interrupt that interrupts that SW itself.

Until now, the models would trigger the interrupt in 1 delta cycle,
which would be after the CPU went to sleep.

With this change, the evaluation of the effect happens immediately
instead.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>