CLOCK (54): Add test interfaces to control duration and to fail tuning
Add a new set of test/cheat interface, to allow controlling how long it
takes for the clocks to start depending on source, how long the
calibration and tunning take, and to pre-set the tunning to fail.

By default they all still take just 0µs, and the tunning never fails.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>